requirements.txt
scholarly>=0.5
elsapy>=0.5.0
python-dotenv>=0.14.0
aiohttp>=3.6.2
beautifulsoup4>=4.9.3
